final changes to code, hopefully

+++ CetinSak/src/experiments/experiment_runner.py
@@ -12,9 +12,9 @@ from src.experiments.model_initialization import model_initialization
 from src.SRHM.optim_loss import loss_func, regularize, opt_algo, measure_accuracy
 from src.SRHM.utils import cpu_state_dict, args2train_test_sizes
 from src.SRHM.observables import locality_measure, state2permutation_stability, state2clustering_error
+
+
 from src.SRHM.diffeomorphism_utilities import *
-
-
 
 from src.SRHM.reference_code.hierarchical import RandomHierarchyModel as RHM
 from src.SRHM.srhm import SparseRandomHierarchyModel as SRHM
@@ -26,6 +26,20 @@ def experiment_run(idx, experiment_name, args):
 
     output_idx = f"outputs/{experiment_name}_{idx}.pkl"
 
+    print(f"Checking if {output_idx} exists.")
+    flag = 0
+    try:
+        with open(output_idx, "r") as handle:
+            flag = 1
+    except:
+        pass
+    if flag == 1:
+        if args.skip_existing:
+            print("")
+            return
+        raise OSError(f"File {output_idx} already exists, do you want to continue training with it?")
+
+    print(f"{output_idx} does not exist, starting training.")
     with open(output_idx, "wb+") as handle:
         pickle.dump(args, handle)
     try:
@@ -33,7 +47,10 @@ def experiment_run(idx, experiment_name, args):
             with open(output_idx, "wb") as handle:
                 pickle.dump(data, handle)
     except:
-        os.remove(output_idx)
+        print("Do you want to delete current results?")
+        confirmation = input("Type 'ok' to delete file : ")
+        if confirmation == 'ok':
+            os.remove(output_idx)
         raise
 
 def run(args):
@@ -63,7 +80,7 @@ def run(args):
         args.batch_size = args.ptr // 2
 
     if args.save_dynamics:
-        dynamics = [{"acc": 0.0, "epoch": 0., "net": cpu_state_dict(net0)}]
+        dynamics = [{"acc": 0.0, "epoch": 0.}]
     else:
         dynamics = None
 
@@ -84,20 +101,20 @@ def run(args):
         epochs_list.append(epoch)
 
         # measuring locality for fcn nets
-        if args.locality == 1:
-            assert args.net == 'fcn', "Locality can only be computed for fcns !!"
-            state = net.state_dict()
-            hidden_layers = [state[k] for k in state if 'w' in k][:-2]
-            with torch.no_grad():
-                locality.append(locality_measure(hidden_layers, args)[0])
+        # if args.locality == 1:
+        #     assert args.net == 'fcn', "Locality can only be computed for fcns !!"
+        #     state = net.state_dict()
+        #     hidden_layers = [state[k] for k in state if 'w' in k][:-2]
+        #     with torch.no_grad():
+        #         locality.append(locality_measure(hidden_layers, args)[0])
 
         # measure stability to semantically equivalent data realizations
-        if args.stability == 1:
-            state = net.state_dict()
-            stability.append(state2permutation_stability(state, args))
-        if args.clustering_error == 1:
-            state = net.state_dict()
-            clustering_error.append(state2clustering_error(state, args))
+        # if args.stability == 1:
+        #     state = net.state_dict()
+        #     stability.append(state2permutation_stability(state, args))
+        # if args.clustering_error == 1:
+        #     state = net.state_dict()
+        #     clustering_error.append(state2clustering_error(state, args))
 
         if epoch % 10 != 0 and not args.save_dynamics: continue
 
@@ -114,13 +131,11 @@ def run(args):
         # ):
             # save dynamics at 30 log-spaced points in time
             dynamics.append(
-                {"acc": acc, "epoch": epoch, "net": cpu_state_dict(net)}
+                {"acc": acc, "epoch": epoch,}
             )
         if acc > best_acc:
             best["acc"] = acc
             best["epoch"] = epoch
-            if args.save_best_net:
-                best["net"] = cpu_state_dict(net)
             # if args.save_dynamics:
             #     dynamics.append(best)
             best_acc = acc
@@ -131,15 +146,13 @@ def run(args):
             "epoch": epochs_list,
             "train loss": loss,
             "terr": terr,
-            "locality": locality,
-            "stability": stability,
-            "clustering_error": clustering_error,
             "dynamics": dynamics,
             "best": best,
         }
 
         if args.seed_diffeo:
             out["acc_diffeo"],out["sensitivity_diffeo"] = test_with_sensitivity(args, testdiffeo, net, criterion, print_flag=epoch % 5 == 0)
+
 
         if args.seed_synonym:
             acc_synonym = test(args, testsynonym, net, criterion, print_flag=epoch % 5 == 0)
@@ -160,40 +173,33 @@ def run(args):
             if trloss_flag >= args.zero_loss_epochs:
                 break
 
-    try:
-        wo = weights_evolution(net0, net)
-    except:
-        print("Weights evolution failed!")
-        wo = None
+    # try:
+    #     wo = weights_evolution(net0, net)
+    # except:
+    #     print("Weights evolution failed!")
+    #     wo = None
 
-    if args.locality == 2:
-        assert args.net == 'fcn', "Locality can only be computed for fcns !!"
-        state = net.state_dict()
-        hidden_layers = [state[k] for k in state if 'w' in k][:-2]
-        with torch.no_grad():
-            locality.append(locality_measure(hidden_layers, args)[0])
+    # if args.locality == 2:
+    #     assert args.net == 'fcn', "Locality can only be computed for fcns !!"
+    #     state = net.state_dict()
+    #     hidden_layers = [state[k] for k in state if 'w' in k][:-2]
+    #     with torch.no_grad():
+    #         locality.append(locality_measure(hidden_layers, args)[0])
 
-    if args.stability == 2:
-        state = net.state_dict()
-        stability.append(state2permutation_stability(state, args))
+    # if args.stability == 2:
+    #     state = net.state_dict()
+    #     stability.append(state2permutation_stability(state, args))
 
-    if args.clustering_error == 2:
-        state = net.state_dict()
-        clustering_error.append(state2clustering_error(state, args))
+    # if args.clustering_error == 2:
+    #     state = net.state_dict()
+    #     clustering_error.append(state2clustering_error(state, args))
 
     out = {
         "args": args,
         "epoch": epochs_list,
         "train loss": loss,
         "terr": terr,
-        "locality": locality,
-        "stability": stability,
-        "clustering_error": clustering_error,
         "dynamics": dynamics,
-        "init": cpu_state_dict(net0) if args.save_init_net else None,
-        "best": best,
-        "last": cpu_state_dict(net) if args.save_last_net else None,
-        "weight_evo": wo,
         'avg_epoch_time': avg_epoch_time,
     }
     yield out
@@ -211,6 +217,10 @@ def train(args, trainloader, net0, criterion):
     num_batches = math.ceil(args.ptr / args.batch_size)
     checkpoint_batches = torch.linspace(0, num_batches, 10, dtype=int)
 
+    if num_batches >= 1000:
+        checkpoint_batches = torch.linspace(0, num_batches, num_batches//100, dtype=int)
+    if num_batches >= 1e6:
+        checkpoint_batches = torch.linspace(0, num_batches, num_batches//1000, dtype=int)
     for epoch in range(args.epochs):
 
         # layerwise training
@@ -224,7 +234,8 @@ def train(args, trainloader, net0, criterion):
         correct = 0
         total = 0
         for batch_idx, (inputs, targets) in enumerate(trainloader):
-
+            if args.net in ['VGG11', 'VGG16', 'ResNet18', 'ResNet34', 'EfficientNetB0']:
+                inputs = inputs.unsqueeze(1)
             inputs, targets = inputs.to(args.device), targets.to(args.device)
             optimizer.zero_grad()
             outputs = net(inputs)
@@ -255,7 +266,6 @@ def train(args, trainloader, net0, criterion):
 
         yield net, epoch + 1, train_loss / (batch_idx + 1), avg_epoch_time
 
-
 def test_with_sensitivity_x(args, testloader, net, criterion, print_flag=True, num_transforms=10,sensitivity_type="diffeo"):
 
     if sensitivity_type == "diffeo":
@@ -275,6 +285,8 @@ def test_with_sensitivity_x(args, testloader, net, criterion, print_flag=True, n
     with torch.no_grad():
         # First pass: collect original outputs
         for batch_idx, (inputs, targets) in enumerate(testloader):
+            if args.net in ['VGG11', 'VGG16', 'ResNet18', 'ResNet34', 'EfficientNetB0']:
+                inputs = inputs.unsqueeze(1)
             inputs, targets = inputs.to(args.device), targets.to(args.device)
             
             outputs = net(inputs)
@@ -326,6 +338,8 @@ def test_with_sensitivity(args, testloader, net, criterion, print_flag=True, num
     with torch.no_grad():
         # First pass: collect original outputs
         for batch_idx, (inputs, targets) in enumerate(testloader):
+            if args.net in ['VGG11', 'VGG16', 'ResNet18', 'ResNet34', 'EfficientNetB0']:
+                inputs = inputs.unsqueeze(1)
             inputs, targets = inputs.to(args.device), targets.to(args.device)
             
             outputs = net(inputs)
@@ -378,6 +392,8 @@ def test(args, testloader, net, criterion, print_flag=True):
     total = 0
     with torch.no_grad():
         for batch_idx, (inputs, targets) in enumerate(testloader):
+            if args.net in ['VGG11', 'VGG16', 'ResNet18', 'ResNet34', 'EfficientNetB0']:
+                inputs = inputs.unsqueeze(1)
             inputs, targets = inputs.to(args.device), targets.to(args.device)
             outputs = net(inputs)
 
@@ -443,11 +459,13 @@ def experiment_initialization(args):
     trainset, testset, input_dim, ch = dataset_initialization(args)
 
     trainloader = torch.utils.data.DataLoader(
-        trainset, batch_size=args.batch_size, shuffle=True, num_workers=0)
+        trainset, batch_size=args.batch_size, shuffle=True, num_workers=2)
+    print("Lenght of trainloader:",len(trainloader))
 
     if testset:
         testloader = torch.utils.data.DataLoader(
-            testset, batch_size=len(testset), shuffle=False, num_workers=0)
+            testset, batch_size=args.batch_size, shuffle=False, num_workers=2)
+        print("Lenght of testloader:",len(testloader))
     else:
         testloader = None
 
@@ -465,7 +483,7 @@ def diffeo_init(args):
     _, testset, _, _ = dataset_initialization(args)
 
     testloader = torch.utils.data.DataLoader(
-        testset, batch_size=args.batch_size, shuffle=False, num_workers=0)
+        testset, batch_size=args.batch_size, shuffle=False, num_workers=4)
 
     args.diffeo_init = False
 
@@ -479,29 +497,17 @@ def synonym_init(args):
     _, testset, _, _ = dataset_initialization(args)
 
     testloader = torch.utils.data.DataLoader(
-        testset, batch_size=args.batch_size, shuffle=False, num_workers=0)
+        testset, batch_size=args.batch_size, shuffle=False, num_workers=4)
     args.synonym_init = True
     return testloader
-
-# Adapted from : https://stackoverflow.com/questions/57416925/best-practices-for-generating-a-random-seeds-to-seed-pytorch
-def get_truly_random_seed_through_os():
-    """
-    Usually the best random sample you could get in any programming language is generated through the operating system. 
-    In Python, you can use the os module.
-
-    source: https://stackoverflow.com/questions/57416925/best-practices-for-generating-a-random-seeds-to-seed-pytorch/57416967#57416967
-    """
-    RAND_SIZE = 4
-    random_data = os.urandom(
-        RAND_SIZE
-    )  # Return a string of size random bytes suitable for cryptographic use.
-    random_seed = int.from_bytes(random_data, byteorder="big")
-    return random_seed
 
 def get_max_size(args):
     d = args.s**args.num_layers
     m_exp = (d-1)/(args.s - 1)
     return args.num_classes * args.m**m_exp
+
+def map_for_nets(x, y):
+    return x.unsqueeze(1), y
 
 def dataset_initialization(args):
     """
@@ -526,66 +532,11 @@ def dataset_initialization(args):
     elif args.dataset == "srhm":
         DataModel = SRHM
 
-
     if args.p and args.p > MAX_SIZE:
-        print("here")
-        concat_count = (args.p // MAX_SIZE) + 1
-        joint_training_loader = DataModel(
-                num_features=args.num_features,
-                m=args.m,  # features multiplicity
-                num_layers=args.num_layers,
-                num_classes=nc,
-                s=args.s, # tuples size
-                s0=args.s0,
-                input_format=args.input_format,
-                whitening=args.whitening,
-                seed=args.seed_init,
-                train=True,
-                transform=transform,
-                testsize=args.pte,
-                max_dataset_size=args.ptr + args.pte,
-                seed_p=seed_p)
-        for _ in concat_count:
-            torch.random_seed(get_truly_random_seed_through_os())
-            new_training_loader = DataModel(
-                num_features=args.num_features,
-                m=args.m,  # features multiplicity
-                num_layers=args.num_layers,
-                num_classes=nc,
-                s=args.s, # tuples size
-                s0=args.s0,
-                input_format=args.input_format,
-                whitening=args.whitening,
-                seed=args.seed_init,
-                train=True,
-                transform=transform,
-                testsize=1,
-                max_dataset_size=1,
-                seed_p=seed_p)
+        print(f"P = {args.p} is larger than MAX_SIZE = {MAX_SIZE}, training will be multiple epochs : {args.p/MAX_SIZE:.2f}")
 
-            joint_training_loader.append(new_training_loader)
-            # TODO : Make joint training set
-
-        trainset = joint_training_loader
-
-        torch.manual_seed(args.seed_init)
-        testset = DataModel(
-                num_features=args.num_features,
-                m=args.m,  # features multiplicity
-                num_layers=args.num_layers,
-                num_classes=nc,
-                s=args.s, # tuples size
-                s0=args.s0,
-                input_format=args.input_format,
-                whitening=args.whitening,
-                seed=args.seed_init,
-                train=False,
-                transform=transform,
-                testsize=args.pte,
-                max_dataset_size=args.ptr+args.pte,
-                seed_p=seed_p)
-
-    else:
+    trainset = None
+    if not (args.diffeo_init or args.synonym_init):
         trainset = DataModel(
             num_features=args.num_features,
             m=args.m,  # features multiplicity
@@ -598,46 +549,48 @@ def dataset_initialization(args):
             seed=args.seed_init,
             train=True,
             transform=transform,
-            testsize=args.pte,
+            testsize=-1,
+            max_dataset_size=args.ptr+args.pte)
+    
+    if args.pte:
+        testset = DataModel(
+            num_features=args.num_features,
+            m=args.m,  # features multiplicity
+            num_layers=args.num_layers,
+            num_classes=nc,
+            s=args.s, # tuples size
+            s0=args.s0,
+            input_format=args.input_format,
+            whitening=args.whitening,
+            seed=args.seed_init,
+            train=False,
+            transform=transform,
+            testsize=-1,
             max_dataset_size=args.ptr+args.pte,
-            seed_p=seed_p)
-        
-        if args.pte:
-            testset = DataModel(
-                num_features=args.num_features,
-                m=args.m,  # features multiplicity
-                num_layers=args.num_layers,
-                num_classes=nc,
-                s=args.s, # tuples size
-                s0=args.s0,
-                input_format=args.input_format,
-                whitening=args.whitening,
-                seed=args.seed_init,
-                train=False,
-                transform=transform,
-                testsize=args.pte,
-                max_dataset_size=args.ptr+args.pte
-            )
-        else:
-            testset = None
-
-    input_dim = trainset[0][0].shape[-1]
-    ch = trainset[0][0].shape[-2] if args.input_format != 'long' else 0
-
-    if args.loss == 'hinge':
-        # change to binary labels
-        trainset.targets = 2 * (torch.as_tensor(trainset.targets) >= nc // 2) - 1
-        if testset:
-            testset.targets = 2 * (torch.as_tensor(testset.targets) >= nc // 2) - 1
-
-    P = len(trainset)
-
-    # take random subset of training set
-    torch.manual_seed(args.seed_trainset)
-    perm = torch.randperm(P)
-    if args.p and args.p > MAX_SIZE:
-        trainset = torch.utils.data.Subset(trainset, perm)
+            seed_p=seed_p
+        )
     else:
+        testset = None
+
+    input_dim, ch = None, None
+    if trainset:
+        input_dim = trainset[0][0].shape[-1]
+        ch = trainset[0][0].shape[-2] if args.input_format != 'long' else 0
+
+        print(f"Input dim : {input_dim}")
+        print(f"ch : {ch}")
+
+        if args.loss == 'hinge':
+            # change to binary labels
+            trainset.targets = 2 * (torch.as_tensor(trainset.targets) >= nc // 2) - 1
+            if testset:
+                testset.targets = 2 * (torch.as_tensor(testset.targets) >= nc // 2) - 1
+
+        P = len(trainset)
+
+        # take random subset of training set
+        torch.manual_seed(args.seed_trainset)
+        perm = torch.randperm(P)
         trainset = torch.utils.data.Subset(trainset, perm[:args.ptr])
 
     return trainset, testset, input_dim, ch
